pyproject.toml
[tool.poetry]
name = "oracle"
version = "1.0.0"
description = "Intraoperative Vital Sign Prediction and Hypotension Onset Detection"
authors = ["Ghanem, Deverett, et al."]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0.0"
numpy = "^1.21.0"
pandas = "^1.4.0"
pyarrow = "^10.0.0"
tables = "^3.7.0"
scikit-learn = "^1.0.0"
matplotlib = "^3.5.0"
pyyaml = "^6.0"
unidecode = "^1.3.0"

[tool.poetry.group.dev.dependencies]
xgboost = "^1.7.0"
wandb = "^0.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
